Move box into a common folder
