+++ js/components/HomeScene/Scene/Common/index.jsx
@@ -1,0 +1,1 @@
+export { default as Box } from "./Box";
